[ADD] exceptions

+++ 4차 LTT/LTT/Controller/RegistrationController.cs
@@ -30,9 +30,18 @@ namespace LTT
             // 1. model에서 curUserID가 관담한거 가져오기
             curUserShoppingBasket = memberRepository.GetUserShoppingBasket(curUserID);
             // 2. view로 보내서 출력하고 수강신청할 과목 받아오기
-            string lectureID = registrationView.RegistrationForm(curUserShoppingBasket);
+            string lectureID = registrationView.RegistrationForm(curUserShoppingBasket, memberRepository.GetUserInfo(curUserID));
             // 3. 진짜 수강신청해주기
-            memberRepository.AddToUserRegistration(curUserID, lectureID);
+            if (ExceptionHandler.CheckIfValidLectureID(lectureID))
+            {
+                memberRepository.AddToUserRegistration(curUserID, lectureID);
+                MyConsole.PrintMessage("수강신청 성공",Console.CursorLeft, Console.CursorTop);
+            }
+            else
+            {
+                MyConsole.PrintMessage("해당 ID는 존재하지 않습니다",Console.CursorLeft, Console.CursorTop);
+            }
+            MyConsole.WaitForEnterKey();
         }
 
         private void GetResult()
@@ -40,7 +49,9 @@ namespace LTT
             // 1. model에서 curUserID가 진짜로 수강신청한거 가죠오기
             curUserRegistrationList = memberRepository.GetUserRegistrationList(curUserID);
             // 2. view로 보내서 출력해주기
-            CommonView.ShowLectureTable(curUserRegistrationList);
+            registrationView.RegistrationResultForm(curUserRegistrationList, memberRepository.GetUserInfo(curUserID));
+            
+            MyConsole.WaitForEnterKey();
         }
 
         private void GetTimeTable()
@@ -52,10 +63,18 @@ namespace LTT
         {
             // 1. model에서 curUserID가 수강신청한거 가져오기
             curUserRegistrationList = memberRepository.GetUserRegistrationList(curUserID);
-            // 2. view로 보내서 출력하고 삭제할 lectureID 받아오기
-            string lectureID = registrationView.RegistrationDeleteForm(curUserRegistrationList);
+            // 2. view로 보내서 진짜로 수강신청한거 출력하고 삭제할 lectureID 받아오기
+            string lectureID = registrationView.RegistrationDeleteForm(curUserRegistrationList, memberRepository.GetUserInfo(curUserID));
             // 3. 수강신청 목록에서 진짜로 삭제하기
-            memberRepository.RemoveFromUserRegistration(curUserID, lectureID);
+            if (ExceptionHandler.CheckIfValidLectureID(lectureID))
+            {
+                memberRepository.RemoveFromUserRegistration(curUserID, lectureID);
+            }
+            else
+            {
+                MyConsole.PrintMessage("THIS LECTURE DOESNT EXIST!", Console.CursorLeft, Console.CursorTop);
+            }
+            MyConsole.WaitForEnterKey();
         }
 
         public void Run()
@@ -70,19 +89,19 @@ namespace LTT
 
                 switch (mode)
                 {
-                    case RegistrationMode.REGST:
+                    case RegistrationMode.REGISTER:
                         Register();
                         break;
 
-                    case RegistrationMode.REGST_RESULT:
+                    case RegistrationMode.REGISTER_RESULT:
                         GetResult();
                         break;
 
-                    case RegistrationMode.REGST_TABLE:
+                    case RegistrationMode.REGISTER_TABLE:
                         GetTimeTable();
                         break;
 
-                    case RegistrationMode.REGST_DELETE:
+                    case RegistrationMode.REGISTER_DELETE:
                         Delete();
                         break;
